+++ CatAnalyzer/test/h2muDraw/h2muDraw_ana_syserr.py
@@ -2,6 +2,7 @@
 import ROOT, CATTools.CatAnalyzer.CMS_lumi, json, os, getopt, sys
 import array
 from CATTools.CatAnalyzer.histoHelper import *
+from CATTools.CatAnalyzer.histoHelper_h2mu import *
 from ROOT import TLorentzVector
 #import DYestimation
 ROOT.gROOT.SetBatch(True)
